tests: Factor out ZulipVerboseEqualTest class, and use more.
